feat: convert pipeline to cli
