+++ scripts/ad_hoc_export.py
@@ -1,9 +1,6 @@
 """Run a one-off export of data."""
 
-from airflow.dag import DagBag
+from app.main import pipeline
 
 if __name__ == "__main__":
-    dagbag = DagBag("airflow")
-    dag = dagbag.get_dag("daily_gads_play_sync")
-    dag.clear()
-    dag.run()
+    pipeline()
